refactor: move VideoPlayer to SharedComponents
