+++ src/components/SharedComponents/VideoPlayer/VideoPlayer.tsx
