+++ src/components/SharedComponents/VideoPlayer/getDuration.test.ts
