+++ src/components/SharedComponents/VideoPlayer/getDuration.ts
