+++ src/components/SharedComponents/VideoPlayer/getPercentageElapsed.ts
