+++ src/components/SharedComponents/VideoPlayer/getTimeElapsed.test.ts
