+++ src/components/SharedComponents/VideoPlayer/getTimeElapsed.ts
