+++ src/components/SharedComponents/VideoPlayer/getVideoThumbnail.ts
