+++ src/components/SharedComponents/VideoPlayer/useSignedVideoToken.ts
@@ -1,10 +1,11 @@
 import { useEffect } from "react";
 import useSWR from "swr";
 
-import errorReporter from "../../common-lib/error-reporter";
-import OakError from "../../errors/OakError";
-
 import getSignedVideoToken from "./getSignedVideoToken";
+
+import errorReporter from "@/common-lib/error-reporter";
+import OakError from "@/errors/OakError";
+
 
 const reportError = errorReporter("useSignedPlaybackId");
 export const apiEndpoint = "/api/video/signed-url";
